multi: move LndNodeSigner into lnd module
