+++ src/lnd.rs
@@ -1,4 +1,11 @@
+use bitcoin::bech32::u5;
+use bitcoin::secp256k1::ecdh::SharedSecret;
+use bitcoin::secp256k1::ecdsa::{RecoverableSignature, Signature};
+use bitcoin::secp256k1::{self, PublicKey, Scalar, Secp256k1};
 use futures::executor::block_on;
+use lightning::chain::keysinterface::{KeyMaterial, NodeSigner, Recipient};
+use lightning::ln::msgs::UnsignedGossipMessage;
+use std::cell::RefCell;
 use std::collections::HashMap;
 use std::path::PathBuf;
 use tonic_lnd::{Client, ConnectError};
@@ -34,4 +41,93 @@ pub(crate) fn features_support_onion_messages(
 ) -> bool {
     features.contains_key(&ONION_MESSAGES_OPTIONAL)
         || features.contains_key(&ONION_MESSAGES_REQUIRED)
+}
+
+pub(crate) struct LndNodeSigner<'a> {
+    pubkey: PublicKey,
+    secp_ctx: Secp256k1<secp256k1::All>,
+    signer: RefCell<&'a mut tonic_lnd::SignerClient>,
+}
+
+impl<'a> LndNodeSigner<'a> {
+    pub(crate) fn new(pubkey: PublicKey, signer: &'a mut tonic_lnd::SignerClient) -> Self {
+        LndNodeSigner {
+            pubkey,
+            secp_ctx: Secp256k1::new(),
+            signer: RefCell::new(signer),
+        }
+    }
+}
+
+impl<'a> NodeSigner for LndNodeSigner<'a> {
+    /// Get node id based on the provided [`Recipient`].
+    ///
+    /// This method must return the same value each time it is called with a given [`Recipient`]
+    /// parameter.
+    ///
+    /// Errors if the [`Recipient`] variant is not supported by the implementation.
+    fn get_node_id(&self, recipient: Recipient) -> Result<PublicKey, ()> {
+        match recipient {
+            Recipient::Node => Ok(self.pubkey),
+            Recipient::PhantomNode => Err(()),
+        }
+    }
+
+    /// Gets the ECDH shared secret of our node secret and `other_key`, multiplying by `tweak` if
+    /// one is provided. Note that this tweak can be applied to `other_key` instead of our node
+    /// secret, though this is less efficient.
+    ///
+    /// Errors if the [`Recipient`] variant is not supported by the implementation.
+    fn ecdh(
+        &self,
+        recipient: Recipient,
+        other_key: &PublicKey,
+        tweak: Option<&Scalar>,
+    ) -> Result<SharedSecret, ()> {
+        match recipient {
+            Recipient::Node => {}
+            Recipient::PhantomNode => return Err(()),
+        }
+
+        // Clone other_key so that we can tweak it (if a tweak is required). We choose to tweak the
+        // `other_key` because LND's API accept a tweak parameter (so we can't tweak our secret).
+        let tweaked_key = if let Some(tweak) = tweak {
+            other_key.mul_tweak(&self.secp_ctx, tweak).map_err(|_| ())?
+        } else {
+            *other_key
+        };
+
+        let shared_secret = match block_on(self.signer.borrow_mut().derive_shared_key(
+            tonic_lnd::signrpc::SharedKeyRequest {
+                ephemeral_pubkey: tweaked_key.serialize().into_iter().collect::<Vec<u8>>(),
+                key_desc: None,
+                ..Default::default()
+            },
+        )) {
+            Ok(shared_key_resp) => shared_key_resp.into_inner().shared_key,
+            Err(_) => return Err(()),
+        };
+
+        match SharedSecret::from_slice(&shared_secret) {
+            Ok(secret) => Ok(secret),
+            Err(_) => Err(()),
+        }
+    }
+
+    fn get_inbound_payment_key_material(&self) -> KeyMaterial {
+        unimplemented!("not required for onion messaging");
+    }
+
+    fn sign_invoice(
+        &self,
+        _hrp_bytes: &[u8],
+        _invoice_data: &[u5],
+        _recipient: Recipient,
+    ) -> Result<RecoverableSignature, ()> {
+        unimplemented!("not required for onion messaging");
+    }
+
+    fn sign_gossip_message(&self, _msg: UnsignedGossipMessage) -> Result<Signature, ()> {
+        unimplemented!("not required for onion messaging");
+    }
 }
